create Appsec::Monitor to handle internal appsec events agnostic to integrations. Move Rack::Reactive::SetUser to Monitor::Reactive::SetUser
