+++ lib/datadog/appsec/contrib/rack/gateway/watcher.rb
@@ -21,7 +21,6 @@ module Datadog
                 watch_request(gateway)
                 watch_response(gateway)
                 watch_request_body(gateway)
-                watch_user_id(gateway)
               end
 
               def watch_request(gateway = Instrumentation.gateway)
@@ -143,49 +142,6 @@ module Datadog
                   next [nil, [[:block, event]]] if block
 
                   ret, res = stack.call(request)
-
-                  if event
-                    res ||= []
-                    res << [:monitor, event]
-                  end
-
-                  [ret, res]
-                end
-              end
-
-              def watch_user_id(gateway = Instrumentation.gateway)
-                gateway.watch('identity.set_user', :appsec) do |stack, user|
-                  block = false
-                  event = nil
-                  waf_context = Datadog::AppSec::Processor.current_context
-
-                  AppSec::Reactive::Operation.new('identity.set_user') do |op|
-                    trace = active_trace
-                    span = active_span
-
-                    Rack::Reactive::SetUser.subscribe(op, waf_context) do |result, _block|
-                      if result.status == :match
-                        # TODO: should this hash be an Event instance instead?
-                        event = {
-                          waf_result: result,
-                          trace: trace,
-                          span: span,
-                          user: user,
-                          actions: result.actions
-                        }
-
-                        span.set_tag('appsec.event', 'true') if span
-
-                        waf_context.events << event
-                      end
-                    end
-
-                    _result, block = Rack::Reactive::SetUser.publish(op, user)
-                  end
-
-                  next [nil, [[:block, event]]] if block
-
-                  ret, res = stack.call(user)
 
                   if event
                     res ||= []
