+++ lib/datadog/appsec/monitor.rb
@@ -1,0 +1,12 @@
+# typed: ignore
+# frozen_string_literal: true
+
+require_relative 'monitor/gateway/watcher'
+
+module Datadog
+  module AppSec
+    # Monitor for internal AppSec Events
+    module Monitor
+    end
+  end
+end
