--- sig/datadog/appsec/contrib/rack/reactive/set_user.rbs
@@ -1,17 +1,0 @@
-module Datadog
-  module AppSec
-    module Contrib
-      module Rack
-        module Reactive
-          module SetUser
-            ADDRESSES: ::Array[::String]
-
-            def self.publish: (untyped op, untyped user) -> untyped
-
-            def self.subscribe: (untyped op, untyped waf_context) { (untyped) -> untyped } -> untyped
-          end
-        end
-      end
-    end
-  end
-end
