+++ sig/datadog/appsec/monitor/gateway/watcher.rbs
@@ -1,0 +1,19 @@
+module Datadog
+  module AppSec
+    module Monitor
+      module Gateway
+        module Watcher
+          def self.watch: () -> untyped
+
+          def self.watch_user_id: (?Datadog::AppSec::Instrumentation::Gateway gateway) -> untyped
+
+          private
+
+          def self.active_trace: () -> (nil | untyped)
+
+          def self.active_span: () -> (nil | untyped)
+        end
+      end
+    end
+  end
+end
